Changes to login.
- models in separate file
- protocols for usecases, datasources and clients
- changes to router
