+++ QuizApp/Login/LoginDatasource.swift
@@ -1,12 +1,17 @@
 import Foundation
 
+protocol LoginDatasourceProtocol {
+
+    func login(username: String, password: String) async throws -> LoginResponseModel
+    
+}
+
 class LoginDatasource {
 
     private var loginClient = LoginClient()
 
     func login(username: String, password: String) async throws -> LoginResponseModel {
-        let response: LoginResponseModel = try await loginClient.login(password: password, username: username)
-        return response
+        return try await loginClient.login(password: password, username: username)
     }
 
 }
