+++ QuizApp/Models/LoginRequestModel.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+struct LoginRequestModel: Encodable {
+
+    let password: String
+    let username: String
+
+}
