+++ QuizApp/Models/LoginResponseModel.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+struct LoginResponseModel: Decodable {
+
+    let accesToken: String
+
+}
+
