+++ QuizApp/Models/RequestError.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+enum RequestError: Error {
+
+    case clientError
+    case serverError
+    case dataError
+    case invalidURL
+    case unknown
+
+}
+
